requirements.txt
flask
pandas
matplotlib
seaborn
python-dotenv
openai
